% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCohort.R
\name{createCohorts}
\alias{createCohorts}
\title{Create the study cohorts}
\usage{
createCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  oracleTempSchema,
  outputFolder
)
}
\arguments{
\item{connectionDetails}{A DatabaseConnector connection-details object.}

\item{cdmDatabaseSchema}{Schema containing the OMOP CDM.}

\item{vocabularyDatabaseSchema}{Schema containing the OMOP vocabulary.}

\item{cohortDatabaseSchema}{Schema where cohorts will be created.}

\item{cohortTable}{Name of the cohort table.}

\item{oracleTempSchema}{Schema used to emulate temporary tables.}

\item{outputFolder}{Folder for cohort counts.}
}
\description{
Create the study cohorts
}
