% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFeatures.R
\name{createFeatures}
\alias{createFeatures}
\title{Create patient-level features}
\usage{
createFeatures(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  oracleTempSchema,
  outputFolder,
  covariateSettingsFile = system.file("settings", "covariateSettings.json", package =
    "SparseMatrix")
)
}
\arguments{
\item{connectionDetails}{A DatabaseConnector connection-details object.}

\item{cdmDatabaseSchema}{Schema containing the OMOP CDM.}

\item{cohortDatabaseSchema}{Schema containing the cohort table.}

\item{cohortTable}{Name of the cohort table.}

\item{oracleTempSchema}{Schema used to emulate temporary tables.}

\item{outputFolder}{Folder for extracted features.}

\item{covariateSettingsFile}{FeatureExtraction settings JSON file.}
}
\description{
Create patient-level features
}
