% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{loadCovariateSettings}
\alias{loadCovariateSettings}
\title{Load FeatureExtraction settings}
\usage{
loadCovariateSettings(file)
}
\arguments{
\item{file}{JSON file created by [saveCovariateSettings()].}
}
\description{
Load FeatureExtraction settings
}
