% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{saveCovariateSettings}
\alias{saveCovariateSettings}
\title{Save FeatureExtraction settings}
\usage{
saveCovariateSettings(covariateSettingsArgs, file)
}
\arguments{
\item{covariateSettingsArgs}{Arguments for
[FeatureExtraction::createCovariateSettings()].}

\item{file}{Destination JSON file.}
}
\description{
Save FeatureExtraction settings
}
