% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema = cohortDatabaseSchema,
  outputFolder,
  createCohorts = TRUE,
  createFeatures = TRUE,
  createSparseMatrix = TRUE,
  covariateSettingsFile = system.file("settings", "covariateSettings.json", package =
    "SparseMatrix"),
  sparseMatrixBatchSize = 1e+06
)
}
\arguments{
\item{connectionDetails}{A DatabaseConnector connection-details object.}

\item{cdmDatabaseSchema}{Schema containing the OMOP CDM.}

\item{cohortDatabaseSchema}{Schema containing the cohort table.}

\item{cohortTable}{Name of the cohort table.}

\item{oracleTempSchema}{Schema used to emulate temporary tables.}

\item{outputFolder}{Folder for logs and results.}

\item{createCohorts}{Create the study cohorts?}

\item{createFeatures}{Extract and save patient-level features?}

\item{createSparseMatrix}{Create the sparse matrix?}

\item{covariateSettingsFile}{FeatureExtraction settings JSON file.}

\item{sparseMatrixBatchSize}{Number of feature rows processed per batch.}
}
\description{
Execute the study
}
